+++ core-service/Data/TypeRegistrationExtensions.cs
@@ -1,0 +1,13 @@
+﻿using CSGOStats.Infrastructure.DataAccess.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace CSGOStats.Services.Core.Data
+{
+    public static class TypeRegistrationExtensions
+    {
+        public static EntityTypeBuilder<TEntity> RegisterTable<TEntity>(this EntityTypeBuilder<TEntity> builder, string schema)
+            where TEntity : class, IEntity =>
+                builder.ToTable(nameof(TEntity), schema);
+    }
+}