+++ core-service/Handling/Entities/IEntityFactory.cs
@@ -1,0 +1,7 @@
+﻿namespace CSGOStats.Services.Core.Handling.Entities
+{
+    public interface IEntityFactory<out TEntity, in TKey>
+    {
+        TEntity CreateEmpty(TKey id);
+    }
+}