+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# v0.1.0
+## Added
+* Startup initialization and services runtime template method.
+* Links extensions
+* Scheduling extensions and abstractions.
+* Handling and core entities abstractions.
+
 # vTemplate
 ## Added
 ## Changed